third_party/WebKit/LayoutTests/inspector/console/console-trace-arguments-expected.txt

CONSOLE MESSAGE: line 9: 1
Tests that console.trace dumps arguments alongside the stack trace.

console-trace-arguments.html:9 1 2 3console-trace-arguments.html:9 a
